requirements.txt
streamlit
anthropic
python-dotenv